Add a RoutingController and tests for routes compatibility.
This simple routing controller is included so that we can set up tests for
Routes functionality. Whether this RoutingController becomes an officially
supported part of TG2.0 or not, it's still important to have tests for the
related Routes functionality, so that we don't break it again.

--HG--
branch : 2.0
extra : rebase_source : 077ab07fe839296873e4ba5fbb15ada33c6cc9a6